Changed default parameters so that sgr can be run out of the box
