+++ splitgraph/config/__init__.py
@@ -21,6 +21,5 @@ PG_PORT = CONFIG["SG_DRIVER_PORT"]
 PG_DB = CONFIG["SG_DRIVER_DB_NAME"]
 PG_USER = CONFIG["SG_DRIVER_USER"]
 PG_PWD = CONFIG["SG_DRIVER_PWD"]
-POSTGRES_CONNECTION = CONFIG["SG_DRIVER_CONNECTION_STRING"]
 SPLITGRAPH_META_SCHEMA = CONFIG["SG_META_SCHEMA"]
 REGISTRY_META_SCHEMA = "registry_meta"
